IMPALA-13809: Update list of excluded locations for the code coverage report
This adds be/src/gutil and be/src/kudu to the list of excluded
locations for the code coverage report. These directories are third-party
party code that have been vendored into the Impala repository. There
is a fair amount of unused code in those directories simply because it
is easier to maintain that way. Impala's tests aren't intending to test
that code.

Testing:
 - Ran code coverage with the updated list

Change-Id: I7f3aa971e50b2c454e9ca607fb9d49d7cc3593ae
Reviewed-on: http://gerrit.cloudera.org:8080/23084
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>