IMPALA-13716 (Addendum): Simplify the call site of createSingleNodePlan()
IMPALA-13716 added an input argument to createSingleNodePlan() that
requires the caller to provide a List of TupleIsNullPredicate. Since
this List depends only on the given PlanNode, this patch adds a
method that computes this List when it is not provided by the caller,
which simplifies the call site as well.

This patch also moves getTupleIsNullPreds() to AnalyticPlanner.java
because it is only used in AnalyticPlanner.java. We make the method
a static one because it's not specific to an instance of
AnalyticPlanner.java.

Change-Id: Ie10361f15db4c0eb2cf5ff5f4d0a28c612b1a930
Reviewed-on: http://gerrit.cloudera.org:8080/22564
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>