IMPALA-13814: test_iceberg_table_metrics fails in non-DFS builds
test_iceberg_table_metrics checks the value of memory-estimate-bytes
for Iceberg tables. The value depends on the number of blocks in the
table which will be 0 in non-DFS (not HDFS, not Ozone) builds. With
this patch the test has different checks for filesystems that support
block location information and for file filesystems that don't.

Change-Id: Ie0c6d8a93e0a07cc680fac6e3f83218b853ca5e3
Reviewed-on: http://gerrit.cloudera.org:8080/22565
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>