IMPALA-13817: Impala fails to start if 'ai_endpoint' and 'ai_additional_platforms' are not set in the right order
The 'ai_endpoint' flag has a validator which depends on
'ai_additional_platforms'. This could result in failure if the
'ai_additional_platforms' is not set before 'ai_endpoint' in the
flagfile.

This patch simplifies the validator for 'ai_endpoint' flag and moves
the supported platforms check to ExecEnv::Init.

Testing:
- Added unit tests for 'is_api_endpoint_supported' function.

Change-Id: I16480564ce434513ab73b531c03e1d229dac2dd6
Reviewed-on: http://gerrit.cloudera.org:8080/22567
Reviewed-by: Yida Wu <wydbaggio000@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>