IMPALA-13738 (Part1): Refactor IcebergTable.getPartialInfo()
This patch aims to refactor Iceberg tables in the Catalog.
It reduces the dependency on the internal hdfsTable_ object in
IcebergTable.getPartialInfo(). Now we depend less on HdfsTable
when we load an Iceberg table in local catalog mode.
The remaining dependencies will be addressed in following patches.

Now creating TPartialPartitionInfo is the responsibility of the
partitions, not the tables:
- FeFsPartition holds a basic default implementation for the
  unpartitioned tables, which includes all Iceberg tables (we
  handle Iceberg tables similarly to unpartitioned HDFS tables
  in the Catalog).
- HdfsPartition extends this implementation with HDFS and
  Hive ACID specific fields.

Testing:
 - existing Iceberg e2e tests in local catalog mode
 - existing Hive ACID tests in local catalog mode
 - ran test_iceberg_with_puffin.py in local catalog mode
 - LocalCatalog FE test

Change-Id: I75d277fb555d0088a2e28111ff529a4605d734fa
Reviewed-on: http://gerrit.cloudera.org:8080/22605
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>