IMPALA-13852: udf-ir.cc missing in ImpalaUdf
IMPALA-11623 separated out *ir.cc files into their own libraries. With
the change, the source file 'be/src/udf/CMakeLists.txt', which was
previously included in the ImpalaUdf library, is now not included. In
case of other libraries, it is intentional that *ir.cc files are not
included in the them as they are linked together with the *Ir libraries
to form the Impala executable. But ImpalaUdf is special because it is
shipped separately as part of the UDF SDK. Because of missing udf-ir.cc,
the functions defined in it, such as
FunctionContext::GetFunctionState(), are not present in libImpalaUdf.a.

This change adds udf-ir.cc to ImpalaUdf.

Testing:
 - verified manually that GetFunctionState() is present in
   libImpalaUdf.a

Change-Id: I7e8ad94931ced21f386e4981f36e0cf7e700ffaf
Reviewed-on: http://gerrit.cloudera.org:8080/22609
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>