IMPALA-13886: Fix order of recursion for complex types for tuple caching
Queries that use complex types fail with an IllegalStateException
when running with tuple caching. The issue is that the logic for
registering tuples is recursing in the wrong order. It processes
the parent before moving on to the child, but the parent references
tuple id's in the child and cannot be processed first. This
switches the ordering so that the children are processed first.

Testing:
 - Added a sanity check for complex types in TupleCacheTest

Change-Id: I853c06e76422382c628850181e3c7e4fdf86efc2
Reviewed-on: http://gerrit.cloudera.org:8080/22667
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>