IMPALA-13891: Unregister MemTrackers from parents in TupleFileReader/Writer
Running failure/test_failpoints.py with tuple caching causes a crash.
The underlying cause is that the MemTrackers for TupleFileReader and
TupleFileWriter get deleted without being unregistered from their
parents, causing a SIGSEGV. Ordinarily, MemTrackers live from Prepare()
to the final teardown of the query. This means that they are always
alive past the point when something would be reading from them.
TupleFileReader/TupleFileWriter can be deleted before the query is
over, so this changes the code to also deregister the MemTracker
from its parent.

Testing:
 - Verified that failure/test_failpoints.py stops crashing when running
   with tuple caching

Change-Id: I14e5484de35f78da74c1246d319553496793de81
Reviewed-on: http://gerrit.cloudera.org:8080/22665
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>