IMPALA-13894: Allow slow check in tuple cache correctness verification when file sizes differ
Currently, tuple cache correctness verification does a fast check,
and returns an error if file sizes are different.

This patch allows a slow check when file sizes differ. Because the
slow check may provide a clearer error message and help prevent
false mismatches when identical rows appear in a different order,
which may lead to size differences. Updated TupleTextFileUtilTest
for this change.

Also fixes argument order in VerifyRows() to correct misleading
log output. The previous order was incorrect, causing the log to
show the wrong one as the reference file.

Tests:
Passed core tests.
Manually verified that when file sizes differ, the query proceeds
to the slow check after this change.

Change-Id: I02e031410dac32d9df746201b156783a8b7d9a1a
Reviewed-on: http://gerrit.cloudera.org:8080/22661
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Kurt Deschler <kdeschle@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>