IMPALA-13914: Revise configuration files of Ranger after RANGER-4445
According to the suggestion by Madhan Neethiraj from Apache Ranger,
this patch revises the configuration files needed by the Ranger server
and the Ranger plug-in so that the Ranger plug-in will still be able to
make REST API calls to the Ranger server in the non-Kerberized
environment after RANGER-4445, where additional calls to
ensureAdminAccess() were added to authenticate the connected plug-in.

Specifically, in the configuration file used by the Ranger server
during startup, we comment out various API endpoints so that the Ranger
server is able to correctly associate a session id with the
corresponding plug-in instance.

In the configuration file needed by the plug-in, to allow the plug-in
to authenticate to the server, we provide the respective username and
password of the user 'admin'. Those 2 configurations were added in
RANGER-4908.

Change-Id: Ie37fa9b5ef26df92496c27bf8a340a54c760c220
Reviewed-on: http://gerrit.cloudera.org:8080/22712
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Fang-Yu Rao <fangyu.rao@cloudera.com>