IMPALA-13873: Missing equivalence conjunct in aggregation node with inline views
Some queries involving plain (distinct) UNIONs miss conjuncts, leading
to incorrect results:

Example:
  WITH u1 AS (select 10 a, 10 b),
  t AS (select a, b, min(b) over (partition by a) min_b from u1 UNION
  select 10, 10, 20)
  select t.* from t where t.b = t.min_b;

Expected result:
  +----+----+-------+
  | a  | b  | min_b |
  +----+----+-------+
  | 10 | 10 | 10    |
  +----+----+-------+

Actual result:
  +----+----+-------+
  | a  | b  | min_b |
  +----+----+-------+
  | 10 | 10 | 10    |
  | 10 | 20 | 10    |
  +----+----+-------+

This is caused by MultiAggregateInfo assuming that conjuncts bound by
grouping slots that are produced by SlotRef grouping expressions are
already evaluated below the AggregationNode. However, this is not true
in all cases: with UNIONs, there may be conjuncts that are unassigned
below the AggregationNode.

This may happen if a conjunct cannot be pushed into all operands of a
UNION, because the source tuples in the operands do not contain all of
the slots referenced by the predicate. In the example above, it happens
in the first operand:
  select a, b, min(b) over (partition by a) min_b from u1
The source tuple, 'u1', contains only two slots ('a' and 'b'), but does
not contain a slot corresponding to 'min(b)' - therefore the predicate
't.b = t.min_b' is not bound by the tuple of 'u1'. In theory, the
predicate could still be evaluated directly after materialising the
tuple with 'min(b)', still inside the UNION operand, but Impala
currently does not work that way.

In these cases, the conjuncts need to be evaluated in the
AggregationNode (possibly in addition to some of the UNION operands).

This change fixes this problem by introducing a method in
MultiAggregateInfo: 'setConjunctsToKeep()', where the caller can pass a
list of conjuncts that will not be eliminated. This is called during the
planning of the UNION if there are unassigned conjuncts remaining.

Testing:
 - Added a PlannerTest and an EE test for the case where a conjunct
   was previously incorrectly removed from the AggregationNode.
 - Existing tests cover the case when conjuncts can be safely removed
   from an AggregationNode above a UnionNode because the conjuncts are
   pushed into all union operands, see for example
   https://github.com/apache/impala/blob/6f2d9a2/testdata/workloads/functional-planner/queries/PlannerTest/union.test#L3914

Change-Id: I67a59cd96d83181ce249fd6ca141906f549a09b3
Reviewed-on: http://gerrit.cloudera.org:8080/22746
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>