IMPALA-13937: Use simpler chmod syntax to set +t on /var/tmp in Docker build
Some Docker base images contain basic Unix utilities implemented by
Busybox instead of the usual linux-coreutils package. The chmod command
in the Busybox implementation seems to ignore certain syntax variants:
the current invocation for setting the sticky bit (+t) on /var/tmp got
silently ignored, while chmod indicated success, returning 0 to the
calling script.

This patch changes the chmod call to a slightly simple syntax, which was
tested to be understood by Busybox and coreutils both; and adds a simple
inline check to assert that the directories required by Kerberos
- exist
- and have the required ownership and permission structure.

The assertion fails the Docker build if setting up /tmp and /var/tmp in
a Kerberos-compatible way did not succeed.

Change-Id: I20c52dc70fb73337efcd6d12652bf99c3c473ff9
Reviewed-on: http://gerrit.cloudera.org:8080/22811
Reviewed-by: Peter Rozsa <prozsa@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>