IMPALA-13968: Fix TestBinaryTypeInText.test_invalid_binary_type in ARM builds
In some platforms there's a bug in libSASL, which makes sasl_decode64()
accept almost anything as input and it won't report any errors.
See details here:
https://github.com/cyrusimap/cyrus-sasl/issues/619

TestBinaryTypeInText::test_invalid_binary_type uses a data file
that has two values that should be Base64 encoded but they aren't.
The test checks that Impala raises the corresponding errors.

The first value doesn't have a correct size, so Impala's
Base64DecodeBufLen() will reject it. The second one passes the
Base64DecodeBufLen() check so we will try to decode it with
Base64Decode() that uses sasl_decode64() under the hood. If
sasl_decode64() has the bug we won't report any errors for the
second value.

This patch changes the second value to one that still passes
the Base64DecodeBufLen() check, but also sasl_decode64() rejects
it in all platforms. To achieve this, the new value uses the
special '=' character.

The bug in sasl_decode64() is also a known issue in Impala,
see IMPALA-9926.

Alternatively, we could create our own Base64 decoder, or use one
from a different library, but it is out of scope of this patch
as it requires more thought and performance measurements.
IMPALA-13973 tracks this option.

Change-Id: Ifff99fd2839e75add49ee835323764615bd5139a
Reviewed-on: http://gerrit.cloudera.org:8080/22791
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>