IMPALA-13974: Don't check catName if fetching COMMIT_COMPACTION_EVENT or ALLOC_WRITE_ID_EVENT
COMMIT_COMPACTION_EVENT is one kind of HMS notification events that are
generated when a transactional table finishes a compaction.
ALLOC_WRITE_ID_EVENT is another kind of transactional table events
generated when a data modification, e.g. INSERT, starts.

Due to the bug of HIVE-28912, the catName of them is always NULL. Note
that catName is an optional field for all kinds of HMS events:
struct NotificationEvent {
    1: required i64 eventId,
    2: required i32 eventTime,
    3: required string eventType,
    4: optional string dbName,
    5: optional string tableName,
    6: required string message,
    7: optional string messageFormat,
    8: optional string catName
}
It's OK to be NULL for some global events like OPEN_TXN, COMMIT_TXN,
ABORT_TXN, etc. But for table level events, catName should be the
catalog name of the table. Usually, it's the default catalog, 'hive'.

When checking COMMIT_COMPACTION_EVENT and ALLOC_WRITE_ID_EVENT events on
a table, we shouldn't set the filter on catName. Otherwise, they will
all be skipped due to this bug.

There are two filters we use in fetching HMS events. One is the server
side filter we set in NotificationEventRequest. The other one is the
client side filter, NotificationFilter, used in filtering events fetched
from HMS.

This patch removes the check on catName in these filters to avoid
missing COMMIT_COMPACTION_EVENT and ALLOC_WRITE_ID_EVENT.

MetastoreEventsProcessorTest.testNotificationEventRequest() has some
wrong expected numbers due to this bug. This patch fixes them and
refactors the code to be more readable.

Tests:
 - Ran test_hms_event_sync_with_commit_compaction 20 times. Without the
   fix, it fails in 2-3 runs.
 - Ran MetastoreEventsProcessorTest.testNotificationEventRequest().

Change-Id: I2e182d32ee2bb8a69c4f71c05eb9e87a5a115f24
Reviewed-on: http://gerrit.cloudera.org:8080/22794
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>